Remove VERSION constant
VERSION constants are an anti-pattern. The only safe way to get the
version of a gem at runtime is through the rubygems gemspec api. Which
means the VERSION constant is only useful as a static file. As such, it
should be in a plaintext or yaml (or somesuch) file as VERSION.